BETA 3.0 b15
Basically reworked the movemsg cmd and it took nearly a whole day so please let this be enough for a commit. You can now move a conversation up to 25 messages to another channel with the **movemsg cmd. Also made some changes to the updating order to make more sense. Minor other changes.