Version 3.0
After over half a year of more or less active work I'm finally releasing Version 3.0. Most of it has been rewritten and there are far too much changes to be able to list all of them here. If you are interested in detailed changelogs feel free to take a look at the commit history for BETA 3.0 pushes on GitHub. So let's see what happens when I push this...